+++ database/migrations/2019_07_30_000021_create_exames_realizados_table.php
@@ -17,9 +17,9 @@ class CreateExamesRealizadosTable extends Migration
             $table->bigIncrements('id');
             $table->unsignedInteger('exame_id');
             $table->foreign('exame_id')->references('id')->on('exames');
-            $table->unsignedInteger('paciente_id');
+            $table->unsignedBigInteger('paciente_id');
             $table->foreign('paciente_id')->references('id')->on('pacientes');
-            $table->unsignedInteger('solicitacao_exame_id');
+            $table->unsignedBigInteger('solicitacao_exame_id');
             $table->foreign('solicitacao_exame_id')->references('id')->on('solicitacoes_exames');
             $table->string('val_resultado', 300);
             $table->timestamps();
